Build pages: root, terms, privacy, newsroom, leadership #153
- add clip path to `quarter-circle` style
- add authored section names to library config
